- tx_report is now 16bit - X0121 will turn on basic AskSin-mode and report decoded messages - EM sending now uses 'M' - command ('E' was taken for Ethernet already)

+++ culfw/clib/rf_receive.h
@@ -29,7 +29,7 @@ uint8_t rf_isreceiving(void);
 uint8_t cksum1(uint8_t s, uint8_t *buf, uint8_t len);
 uint8_t cksum2(uint8_t *buf, uint8_t len);
 
-extern uint8_t tx_report;
+extern uint16_t tx_report;
 
 void RfAnalyze_Task(void);
